requirements.txt
websockets==13.1
pygame
pyaudio
python-dotenv
numpy
browser_use
playwright